Extract logic to DocumentLoadError component
